requirements.txt
fastapi[standard]
python-dotenv
langchain
langchain-community
openai
langchain-openai
pymongo